Base da API finalizada

+++ BazarCarioca.WebAPI/Controllers/ProductTypesController.cs
@@ -1,60 +1,84 @@
-﻿using BazarCarioca.WebAPI.Models;
+﻿using AutoMapper;
+using BazarCarioca.WebAPI.DTOs;
+using BazarCarioca.WebAPI.Extensions;
+using BazarCarioca.WebAPI.Models;
 using BazarCarioca.WebAPI.Repositories;
+using Microsoft.AspNetCore.JsonPatch;
 using Microsoft.AspNetCore.Mvc;
+using Newtonsoft.Json;
 
 namespace BazarCarioca.WebAPI.Controllers
 {
-    [Route("BazarCarioca/Tipos_de_Produtos")]
+    [Route("Bazar-Carioca/Tipos-de-Produtos")]
     [ApiController]
     public class ProductTypeController : ControllerBase
     {
         private readonly IProductTypeRepository Repository;
+        private readonly IMapper Mapper;
 
-        public ProductTypeController(IProductTypeRepository repository)
+        public ProductTypeController(IProductTypeRepository _Repository, IMapper mapper)
         {
-            Repository = repository;
+            Repository = _Repository;
+            Mapper = mapper;
         }
 
         [HttpGet]
         public async Task<IActionResult> Get()
         {
             var productTypes = await Repository.GetAsync();
-
+            if (productTypes.IsNullOrEmpty())
+                return NotFound("Nenhum tipo de produto foi encontrado.");
             return Ok(productTypes);
         }
 
         [HttpGet("{Id:int}")]
-        public async Task<ActionResult<ProductType>> GetById(int Id)
+        public async Task<IActionResult> GetById(int Id)
         {
-            var productTypes = await Repository.GetByIdAsync(Id);
-
-            return Ok(productTypes);
+            var productType = await Repository.GetByIdAsync(Id);
+            if (productType == null)
+                return NotFound($"O tipo de produto com Id = {Id} não foi encontrado.");
+            return Ok(productType);
         }
 
-        [HttpPost("Criar")]
-        public async Task<ActionResult<ProductType>> CreateStore([FromBody] ProductType productType)
+        [HttpPost]
+        public async Task<IActionResult> Create([FromForm] ProductTypeDTO createDto)
         {
+            if (createDto == null)
+                return BadRequest("Nenhuma informação foi passada na requisição.");
+
+            var productType = Mapper.Map<ProductType>(createDto);
             await Repository.AddAsync(productType);
 
-            return Ok(productType);
+            var productTypeDto = Mapper.Map<ProductTypeDTO>(productType);
+            return Ok(productTypeDto);
         }
 
-        [HttpPut("Atualizar/{Id:int}")]
-        public async Task<ActionResult<ProductType>> FullUpdate(int Id, [FromBody] ProductType productType)
+        [HttpPatch("{Id:int}")]
+        [Consumes("multipart/form-data")]
+        public async Task<IActionResult> Patch(int Id, [FromForm] string requestJson)
         {
-            productType.Id = Id;
-            await Repository.UpdateAsync(Id, productType);
+            if (requestJson == null)
+                return BadRequest("Houve um erro na requisição HTTP. Informações não foram enviadas.");
 
-            return Ok(productType);
+            var request = JsonConvert.DeserializeObject<JsonPatchDocument<ProductType>>(requestJson);
+            var productType = await Repository.GetByIdAsync(Id);
+            if (productType == null)
+                return BadRequest($"Não existe um tipo de produto com o Id = {Id} para ser alterado.");
+
+            var patchedProductType = await Repository.UpdateAsync(productType, request);
+            var productTypeDto = Mapper.Map<ProductTypeDTO>(productType);
+            return Ok(productTypeDto);
         }
 
-        [HttpDelete("Apagar/{Id:int}")]
-        public async Task<ActionResult<bool>> DeleteStore(int Id)
+        [HttpDelete("{Id:int}")]
+        public async Task<IActionResult> Delete(int Id)
         {
+            var productType = await Repository.GetByIdAsync(Id);
+            if (productType == null)
+                return NotFound($"O tipo de produto não foi apagado pois não existe um tipo de produto com Id = {Id}.");
+
             await Repository.DeleteAsync(Id);
-
-            return Ok($"Tipo de produto com id = {Id} apagado.");
+            return Ok($"Tipo de produto com id = {Id} foi apagado.");
         }
     }
 }
-
